chore(core-manager): setEnv accepts object in parameters
